+++ app/data/spells/air.ts
@@ -1,0 +1,112 @@
+import { Spell } from "../../types/spell.ts";
+
+export const airSpell: Array<Spell> = [
+  {
+    id: "34",
+    name: "Haste",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "35",
+    name: "View Air",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "36",
+    name: "Disguise",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "37",
+    name: "Disrupting Ray",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "38",
+    name: "Fortune",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "39",
+    name: "Lightning Bolt",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "40",
+    name: "Precision",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "41",
+    name: "Protection from Air",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "42",
+    name: "Air Shield",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "43",
+    name: "Destroy Undead",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "44",
+    name: "Hypnotize",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "45",
+    name: "Chain Lightning",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "46",
+    name: "Counterstrike",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "47",
+    name: "Dimension Door",
+    level: 5,
+    effect: "",
+  },
+  {
+    id: "48",
+    name: "Fly",
+    level: 5,
+    effect: "",
+  },
+  {
+    id: "49",
+    name: "Magic Mirror",
+    level: 5,
+    effect: "",
+  },
+  {
+    id: "50",
+    name: "Summon Air Elemental",
+    level: 5,
+    effect: "",
+  },
+  {
+    id: "51",
+    name: "Titan's Lightning Bolt",
+    level: 5,
+    effect: "",
+  },
+];