+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add fire spell end point.
 - Add water spell end point.
+- Add air spell end point.
 
 #v0.1.0